requirements.txt
python-telegram-bot==21.5
anthropic==0.39.0
